requirements.txt
aioice==0.9.0
aiortc==1.8.0
anyio==4.3.0
astmonkey==0.3.6
asyncclick==8.1.7.2
attrs==23.2.0
av==11.0.0
behave==1.2.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorzero==2.0
coverage==7.4.4
cryptography==42.0.5
dnspython==2.6.1
google-crc32c==1.5.0
gpiozero==1.6.2
idna==3.6
ifaddr==0.2.0
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mock==4.0.3
MutPy==0.6.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pluggy==1.4.0
py==1.11.0
PyAudio==0.2.14
pycparser==2.22
pydot==2.0.0
pyee==11.1.0
pygame==2.5.2
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==6.2.5
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.1
termcolor==2.4.0
toml==0.10.2
typing_extensions==4.10.0
urllib3==2.2.1
tensorflow
numpy 
openai
pynmea2
gtts
gtts-token
adafruit-circuitpython-dht
luma.oled
#picamera2
